JOE-851: Split globals.css and surface-styles by domain
Break the 3.6k-line app globals and 2.4k-line UI surface CSS into
owned domain modules (shell/chat/studio/settings/base + per-surface
UI styles). Keep barrels for renderer entry and document ownership
in design-system.md. Update CSS parity/regression readers for the split.
