+++ tests/design-tokens-sync.test.ts
@@ -22,7 +22,10 @@ function namedBlock(source: string, name: string) {
 }
 
 function desktopTokenDeclarations() {
-  const globals = readFileSync('packages/app/src/styles/globals.css', 'utf8')
+  const globals = [
+    readFileSync('packages/app/src/styles/globals.css', 'utf8'),
+    readFileSync('packages/app/src/styles/domains/base.css', 'utf8'),
+  ].join('\n')
   const generated = readFileSync('packages/app/src/styles/generated/design-tokens.css', 'utf8')
   return {
     generated,
